KSK-1 : Remove unused code, move KoskiConstants to KoskiUtil
